+++ tests/surfaces/test_extrusion_surface3d.py
@@ -157,6 +157,14 @@ class TestExtrusionSurface3D(unittest.TestCase):
         self.assertTrue(contour2d.is_ordered())
         self.assertAlmostEqual(contour2d.area(), 0.012120134592666365, 2)
 
+        surface = surfaces.ExtrusionSurface3D.load_from_file(
+            os.path.join(folder, "periodical_extrusionsurface_linesegment3d_to_2d.json"))
+        contour = vmw.Contour3D.load_from_file(
+            os.path.join(folder, "periodical_extrusionsurface_linesegment3d_to_2d_contour.json"))
+        contour2d = surface.contour3d_to_2d(contour)
+        self.assertTrue(contour2d.is_ordered(1e-5))
+        self.assertAlmostEqual(contour2d.area(), 0.007376809172328507, 2)
+
 
 if __name__ == '__main__':
     unittest.main()